Implement generateCollectionType in cpp
