+++ CMakeLists.txt
@@ -111,7 +111,8 @@ ENDIF(NOT SPLASH_RELEASE)
 SET(SPLASH_LIBS z ${HDF5_LIBRARIES})
 
 # serial or parallel version of libSplash
-SET(SPLASH_CLASSES logging DCAttribute DCDataSet DCGroup HandleMgr SerialDataCollector DomainCollector SDCHelper)
+SET(SPLASH_CLASSES logging DCAttribute DCDataSet DCGroup HandleMgr SerialDataCollector
+    DomainCollector SDCHelper generateCollectionType)
 IF(HDF5_IS_PARALLEL)
     #parallel version
     MESSAGE(STATUS "Parallel HDF5 found. Building parallel version")
